Calculate Quarterly Days in AR

Calculate Quarterly Days in AR

Usage

dar_qtr(df, date = date, gct = gct, earb = earb, dart = 35)

Arguments

df

data frame containing at least three columns: a date column, a gross charges column and an ending AR column

date

column containing a date within the month that Days in AR is to be calculated

gct

column containing a month's total Gross Charges

earb

column containing a month's Ending AR balance

dart

target Days in AR, default is 35 days

Value

data frame

Examples

dar_mon_ex <- data.frame(
date = as.Date(c(
"2022-01-01", "2022-02-01", "2022-03-01",
"2022-04-01", "2022-05-01", "2022-06-01",
"2022-07-01", "2022-08-01", "2022-09-01",
"2022-10-01", "2022-11-01", "2022-12-01")),

gct = c(
325982.23, 297731.74, 198655.14,
186047.56, 123654.34, 131440.28,
153991.95, 156975.52, 146878.12,
163799.44, 151410.74, 169094.46),

earb = c(
288432.52, 307871.08, 253976.56,
183684.92, 204227.59, 203460.47,
182771.32, 169633.64, 179347.72,
178051.11, 162757.49, 199849.32))

dar_qtr(dar_mon_ex, date, gct, earb, dart = 40)
#>         date nqtr ndip  gct_qtr     earb earb_trg    earb_dc earb_pct     adc
#> 1 2022-03-01    1   90 822369.1 253976.6 365497.4 -111520.82   -43.91 9137.43
#> 2 2022-06-01    2   91 441142.2 203460.5 193908.6    9551.82     4.69 4847.72
#> 3 2022-09-01    3   92 457845.6 179347.7 199063.3  -19715.58   -10.99 4976.58
#> 4 2022-12-01    4   92 484304.6 199849.3 210567.2  -10717.91    -5.36 5264.18
#>     dar  pass actual ideal radiff
#> 1 27.80  TRUE   0.31  0.44  -0.13
#> 2 41.97 FALSE   0.46  0.44   0.02
#> 3 36.04  TRUE   0.39  0.43  -0.04
#> 4 37.96  TRUE   0.41  0.43  -0.02