Calculate Monthly Days in AR

Calculate Monthly Days in AR

Usage

dar_month(df, date, gct, earb, dart = 35)

Arguments

df

data frame containing at least three columns: a date column, a gross charges column and an ending AR column

date

column containing a date within the month that Days in AR is to be calculated

gct

column containing a month's total Gross Charges

earb

column containing a month's Ending AR balance

dart

target Days in AR, default is 35 days

Value

data frame

Examples

dar_mon_ex <- data.frame(
date = as.Date(c(
"2022-01-01", "2022-02-01", "2022-03-01",
"2022-04-01", "2022-05-01", "2022-06-01",
"2022-07-01", "2022-08-01", "2022-09-01",
"2022-10-01", "2022-11-01", "2022-12-01")),

gct = c(
325982.23, 297731.74, 198655.14,
186047.56, 123654.34, 131440.28,
153991.95, 156975.52, 146878.12,
163799.44, 151410.74, 169094.46),

earb = c(
288432.52, 307871.08, 253976.56,
183684.92, 204227.59, 203460.47,
182771.32, 169633.64, 179347.72,
178051.11, 162757.49, 199849.32))

dar_month(dar_mon_ex, date, gct, earb, dart = 40)
#>          date     month nmon ndip      gct     earb earb_trg     earb_dc
#> 1  2022-01-01   January    1   31 325982.2 288432.5 420622.2 -132189.712
#> 2  2022-02-01  February    2   28 297731.7 307871.1 425331.1 -117459.977
#> 3  2022-03-01     March    3   31 198655.1 253976.6 256329.2   -2352.653
#> 4  2022-04-01     April    4   30 186047.6 183684.9 248063.4  -64378.493
#> 5  2022-05-01       May    5   31 123654.3 204227.6 159554.0   44673.603
#> 6  2022-06-01      June    6   30 131440.3 203460.5 175253.7   28206.763
#> 7  2022-07-01      July    7   31 153992.0 182771.3 198699.3  -15927.970
#> 8  2022-08-01    August    8   31 156975.5 169633.6 202549.1  -32915.418
#> 9  2022-09-01 September    9   30 146878.1 179347.7 195837.5  -16489.773
#> 10 2022-10-01   October   10   31 163799.4 178051.1 211354.1  -33303.006
#> 11 2022-11-01  November   11   30 151410.7 162757.5 201881.0  -39123.497
#> 12 2022-12-01  December   12   31 169094.5 199849.3 218186.4  -18337.080
#>       earb_pct       adc      dar  pass    actual    ideal     radiff
#> 1  -45.8303773 10515.556 27.42913  TRUE 0.8848106 1.290323 -0.4055120
#> 2  -38.1523257 10633.276 28.95355  TRUE 1.0340553 1.428571 -0.3945161
#> 3   -0.9263268  6408.230 39.63287  TRUE 1.2784797 1.290323 -0.0118429
#> 4  -35.0483280  6201.585 29.61903  TRUE 0.9873009 1.333333 -0.3460325
#> 5   21.8744210  3988.850 51.19962 FALSE 1.6516007 1.290323  0.3612781
#> 6   13.8635104  4381.343 46.43793 FALSE 1.5479309 1.333333  0.2145976
#> 7   -8.7146990  4967.482 36.79355  TRUE 1.1868888 1.290323 -0.1034338
#> 8  -19.4038270  5063.726 33.49976  TRUE 1.0806375 1.290323 -0.2096850
#> 9   -9.1943033  4895.937 36.63195  TRUE 1.2210649 1.333333 -0.1122684
#> 10 -18.7041834  5283.853 33.69721  TRUE 1.0870068 1.290323 -0.2033158
#> 11 -24.0379086  5047.025 32.24821  TRUE 1.0749402 1.333333 -0.2583931
#> 12  -9.1754528  5454.660 36.63827  TRUE 1.1818798 1.290323 -0.1084428